WIP - Partially implemented max updates
Deletes many trivial "pass-through" tests.

Using Integration tests instead of unit tests in all cases (for now).

Trying to use an inMemoryDB, since a good portion of the logic happens
in the repo right now.

Fighting through some typing weirdness with typegoose. Think this
implementation is close to the answer, but trying to find a good
portable way to always have _id on the schema types.